Separate the analysis framework in a different file.
