+++ src/ir/context.rs
@@ -5,7 +5,8 @@ use super::int::IntKind;
 use super::item::{IsOpaque, Item, ItemAncestors, ItemCanonicalPath, ItemSet};
 use super::item_kind::ItemKind;
 use super::module::{Module, ModuleKind};
-use super::named::{UsedTemplateParameters, analyze};
+use super::named::UsedTemplateParameters;
+use super::analysis::analyze;
 use super::template::{TemplateInstantiation, TemplateParameters};
 use super::traversal::{self, Edge, ItemTraversal};
 use super::ty::{FloatKind, Type, TypeKind};
